init.d脚本文件改为软链接；优化启动流程；日志时间显示

+++ shellinabox/shellinabox/scripts/shellinabox_start.sh
@@ -4,30 +4,26 @@ eval `dbus export webshell_`
 source /jffs/softcenter/scripts/base.sh
 
 stop_webshell() {
-killall -9 shellinaboxd
- echo "webshell已关闭。">> /tmp/webshell.log
-[ -e "/jffs/softcenter/init.d/S96webshell.sh" ] && rm -rf /jffs/softcenter/init.d/S96webshell.sh
+if [ -n "$(pidof shellinaboxd)" ]; then
+    killall -9 shellinaboxd
+    echo "【$(date +%Y年%m月%d日\ %X)】: webshell已关闭。">> /tmp/webshell.log
+fi
 }
 
 start_webshell() {
-[ "$webshell_enable" == "0" ] && exit 0
-#不重复启动
-icount=`ps -w|grep shellinaboxd|grep -v grep|wc -l`
-if [ $icount != 0  ] ;then
-stop
-sleep 2
+if [ "$webshell_enable" != "1" ];then
+    [ -L "/jffs/softcenter/init.d/S96webshell.sh" ] && rm -f /jffs/softcenter/init.d/S96webshell.sh
+    exit 0
 fi
 /jffs/softcenter/shellinabox/shellinaboxd --css=/jffs/softcenter/shellinabox/white-on-black.css --service=/:LOGIN -b --disable-ssl
- echo "webshell已启动。">> /tmp/webshell.log
-[ ! -e "/jffs/softcenter/init.d/S96webshell.sh" ] && cp -rf /jffs/softcenter/scripts/shellinabox_start.sh /jffs/softcenter/init.d/S96webshell.sh
+echo "【$(date +%Y年%m月%d日\ %X)】: webshell已启动。">> /tmp/webshell.log
+[ ! -L "/jffs/softcenter/init.d/S96webshell.sh" ] && ln -sf /jffs/softcenter/scripts/shellinabox_start.sh /jffs/softcenter/init.d/S96webshell.sh
 }
 
 restart_webshell() {
   stop_webshell
   sleep 1
-  if [ "$webshell_enable" == "1" ] ;then
   start_webshell
-  fi
 }
 
 case $ACTION in
